Change material runtime accumulator to double for better precision
Change material runtime accumulator to double for better precision.

+++ graphics/material.cpp
@@ -2451,7 +2451,7 @@ std::string TextMaterial::GetClassId() const
 { return {}; }
 void TextMaterial::Update(float dt)
 {}
-void TextMaterial::SetRuntime(float runtime)
+void TextMaterial::SetRuntime(double runtime)
 {}
 void TextMaterial::SetUniform(const std::string& name, const Uniform& value)
 {}
